+++ libs/pymod/src/mod/py/causality/EventTrace.cpp
@@ -7,8 +7,6 @@
 // rst:
 // rst:	.. versionadded:: 1.1
 // rst:
-// rst:	.. todo:: check version added before release
-// rst:
 namespace mod::causality::Py {
 namespace {