+++ libs/pymod/src/mod/py/causality/Petri.cpp
@@ -15,8 +15,6 @@ void Petri_doExport() {
 	// rst:
 	// rst:		.. versionadded:: 1.1
 	// rst:
-	// rst:		.. todo:: check version added before release
-	// rst:
 	py::class_<Net, std::shared_ptr<Net>, boost::noncopyable>("Net", py::no_init)
 			// rst:		.. method:: __init__(dg)
 			// rst:
@@ -55,8 +53,6 @@ void Petri_doExport() {
 	// rst:		methods on the object.
 	// rst:
 	// rst:		.. versionadded:: 1.1
-	// rst:
-	// rst:		.. todo:: check version added before release
 	// rst:
 	py::class_<Marking>("Marking", py::no_init)
 			// rst:		.. method:: __init__(net)